+++ hnbm/__init__.py
@@ -1,5 +1,5 @@
 from .estimator import HNBM, HNBMClassifier, HNBMRegressor
-from .losses import Logistic, MeanSquaredError, PseudoHuber, Quantile
+from .losses import Logistic, MeanSquaredError, PseudoHuber, Quantile, Softmax
 from .nn_boost import NNBoost, NNBoostClassifier, NNBoostRegressor
 from .nn_learner import ShallowNNRegressor, make_shallow_nn_pool
 
@@ -11,10 +11,11 @@ __all__ = [
     "MeanSquaredError",
     "PseudoHuber",
     "Quantile",
+    "Softmax",
     "NNBoost",
     "NNBoostClassifier",
     "NNBoostRegressor",
     "ShallowNNRegressor",
     "make_shallow_nn_pool",
 ]
-__version__ = "1.0.1"
+__version__ = "1.2.0"